Módulo_31_-_Streamlit_V/ex040/requirements.txt
streamlit==1.23.1
pandas==1.3.5
numpy==1.21.6
XlsxWriter==3.1.9
openpyxl==3.0.10
pillow==8.4.0